requirements.txt
pyyaml>=6.0
requests>=2.25.0
setuptools>=42.0.0
